Extract action review projection helpers
